requirements.txt
fastapi
uvicorn
python-telegram-bot
python-dotenv
bs4
requests
scrapy
scrapy-random-useragent